refactor(architure): refactor GUI and entity code for transitioning away from ECS but stay close to DoD
Reorganized the project by replacing standalone GUI and entities with an DOD-based structure for future scalability. The goal is to avoid the complexity of ECS but stay in a DoD architecture using SoA.

Introduced new managers, systems and collections for GUI components and particles while removing outdated entity-specific systems.

The TitleScreen is functional again except the Mouse events are not properly processed. This will be the next step.